Immediately show next chord after correct answer
After submitting the correct answer, the user had to press the button
another time to load the next exercise. This was perceived as annoying
and useless. Since the answer is correct, the only feedback needed is
that the answer was correct. The green mark below the exercise is
enough to provide this feedback about the answer being correct.
s