Add support for compound intervals to interval() function
The octave interval was needed for the interval exercise.
Adding support also for intervals up to 14th required
approximately the same amount of work, and it's better
for future developments (it's likely that those intervals
will be needed). This was achieved by creating a raiseOctave()
function, which raises its input abc notation up an octave.
All the functionality of raiseOctave wasn't immediately
needed, but since it required only a small amount of additional
work it was done like this, so that it can be useful also for
other possible future developments. With raiseOctave(), interval()
required only changes in a couple of rows.