Remove unwritable intervals and test that those remaining are writable
Some available intervals couldn't be written for some available roots;
for example Db diminished sixth has the same pitch as Ab, but cannot
be written as B with an accidental (the lowest altered B would be Bx,
which is the same pitch as A, and a half tone more than the desired note).

A test controls that all the remaining available interval/root
combinations are writable.