Add checks for pointsAwarded
If the student has already completed the exercise, a green check sign
is always displayed in the top right corner. When loading the page,
a completed exercise renders the complete screen with "practice more"
("harjoittele lisää" in Finnish) but without the progress bar.
When the student completes an exercise, the complete screen includes
the progress bar.