Fix augmented unison notation bug
The notation() for augmented unison didn't put the natural
accidental when needed, so that the notation and the sound
were those of a perfect unison. Now each notation() function
in chords, intervals and scales uses concatenate for every
note, including the root, so that the natural symbol is
added when needed.