Change size of exercise set user for testing
Everybody was always using the exercise set with size
one for testing, and this has led to countless bugs,
because it tests only the case of the last exercise.
Now size is set to 2, so it's still short, but it
tests also the case of right answer but not-last
exercise.