Inform user if quiz data was not received
If the request for quiz data doesn't get a response
from the server, quizItemId is not known, thus the user
can't get points once the exercise is completed. Now
the user is informed that quiz data wasn't received,
and can decide to reload the page if interested in
getting points registered when the exercise is completed.