Add limits to notes on score
Now the user can't submit an answer unless a
specified number of notes has been written on the
score. It's also possible to limit the maximum
amount of notes on the score to a desired value.
Min and max are customizabile, depending on the
type of piano exercise (Chord, Interval or Scale).