Fix rendering of correct notes in piano exercises
Sometimes when the user pressed a note which was part of the correct
answer, that note was still rendered as an incorrect enharmonic; for
example, in Ab natural minor scale, Cb was rendered as B.

Before, the correct notation was searched from within a string
corresponding to the notation of the correct answer, which is
returned from concatenateNotes. In the previous example, when the
user presses the B key on the piano, chooseFromNoteOptions searches
"B" into the correct notation string, which is "_A_B_c_d_e_f_g_a",
finds "B" and accepts it as the correct alternative, while the "B"
in the correct notation string is actually part of the notation for
the note "_B".

Now, the correct notation is searched from within a list of strings,
in which each string corresponds to the notation for one note. So
now "B" isn't found into the list (the list includes "_B", not "B"),
and the correct choice "_C" is selected.

There is still the problem of notes being sometimes rendered in the
wrong octave.