Fix wrong octave note rendering for Cb and B#
In the piano exercises Cb and B# were rendered in the wrong octave.
This has been resolved by modifying the notes array in roots, so
that each subarray contains the notation for the central octave.
Then the octave logic in the interval function had to be modified
as a result, and also the octave logic in the pianoToNotation file.
This required a little refactoring in the PianoExercise files, too.